requirements.txt
absl-py==0.13.0
attrs==21.2.0
cachetools==4.2.2
click==8.0.1
click-plugins==1.1.1
cycler==0.10.0
dill==0.3.4
future==0.18.2
glcontext==2.3.4
iso8601==0.1.16
kiwisolver==1.3.2
matplotlib==3.4.3
moderngl==5.6.4
multiprocess==0.70.12.2
numpy==1.21.2
opencv-python==4.5.3.56
ortools==9.0.9048
penkit-optimize==0.0.3
Pillow==8.3.2
pnoise==0.1.0
protobuf==3.18.0
pyparsing==2.4.7
pyserial==3.5
PySide2==5.15.2
python-dateutil==2.8.2
PyYAML==5.4.1
Rtree==0.9.7
scipy==1.7.1
Shapely==1.7.1
shiboken2==5.15.2
six==1.16.0
svgelements==1.5.4
svgpathtools==1.4.1
svgwrite==1.4.1
toml==0.10.2
tqdm==4.62.2
vpype==1.7.0
vpype-gcode==0.9.0
